# Optimize to_rows