The previous implementation, which relied on Enumerable,
was droping one column at a time, which meant the dataframe
was copied once per column.

The new version removes the unused enumerable implementation
and traverses the names instead.