@joshuataylor 2022-06-22
Before:
```
Operating System: Linux
CPU Information: AMD Ryzen 5 5600X 6-Core Processor
Number of Available Cores: 12
Available memory: 31.26 GB
Elixir 1.13.4
Erlang 25.0.1

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 30 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 32 s

Benchmarking to_rows before ...

Name                     ips        average  deviation         median         99th %
to_rows before          1.63      613.96 ms    ±13.65%      634.19 ms      720.78 ms
```

After:
```
==> explorer
Compiling 2 files (.ex)
Generated explorer app
Operating System: Linux
CPU Information: AMD Ryzen 5 5600X 6-Core Processor
Number of Available Cores: 12
Available memory: 31.26 GB
Elixir 1.13.4
Erlang 25.0.1

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 30 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 32 s

Benchmarking to_rows before ...

Name                     ips        average  deviation         median         99th %
to_rows after          1.66      602.44 ms     ±7.70%      608.86 ms      696.37 ms
```

Removing the zip `|> Enum.zip_with(fn row -> keys |> Enum.zip(row) |> Map.new() end)`:

Before:
```
Operating System: Linux
CPU Information: AMD Ryzen 5 5600X 6-Core Processor
Number of Available Cores: 12
Available memory: 31.26 GB
Elixir 1.13.4
Erlang 25.0.1

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 30 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 32 s

Benchmarking to_rows before ...

Name                     ips        average  deviation         median         99th %
to_rows before          2.84      351.53 ms    ±11.78%      377.99 ms      418.10 ms
```

After
```
Operating System: Linux
CPU Information: AMD Ryzen 5 5600X 6-Core Processor
Number of Available Cores: 12
Available memory: 31.26 GB
Elixir 1.13.4
Erlang 25.0.1

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 30 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 32 s

Benchmarking to_rows before ...

Name                     ips        average  deviation         median         99th %
to_rows before          3.23      309.35 ms     ±3.85%      307.52 ms      386.00 ms
```