@josevalim 2022-06-22
Thanks! I was expecting to see mostly a reduction on memory usage, could you see anything like that?